LayoutTests/platform/mac/accessibility/button-shouldnot-have-axvalue.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<input tabindex=0 id="button" type="button" value="test">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that a button should not have AXRequired and should not have AXValue returned.");

    if (window.accessibilityController) {

          document.getElementById("button").focus();
          var button = accessibilityController.focusedElement;

          shouldBe("button.isAttributeSupported('AXValue')", "false");
          shouldBe("button.isAttributeSupported('AXRequired')", "false");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
